+++ app/desmos/cmd/root.go
@@ -5,15 +5,18 @@ import (
 	"os"
 	"path/filepath"
 
+	"github.com/desmos-labs/desmos/app/desmos/cmd/chainlink"
+	"github.com/desmos-labs/desmos/app/desmos/cmd/sign"
+
 	config "github.com/cosmos/cosmos-sdk/client/config"
 
 	"github.com/cosmos/cosmos-sdk/x/crisis"
 
-	"github.com/desmos-labs/desmos/app"
-
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/simapp/params"
 	"github.com/cosmos/cosmos-sdk/snapshots"
+
+	"github.com/desmos-labs/desmos/app"
 
 	"github.com/spf13/cast"
 	"github.com/spf13/cobra"
@@ -34,7 +37,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authclient "github.com/cosmos/cosmos-sdk/x/auth/client"
 	authcmd "github.com/cosmos/cosmos-sdk/x/auth/client/cli"
-	"github.com/cosmos/cosmos-sdk/x/auth/types"
+	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	cosmosgenutilcli "github.com/cosmos/cosmos-sdk/x/genutil/client/cli"
 
@@ -56,7 +59,7 @@ func NewRootCmd() (*cobra.Command, params.EncodingConfig) {
 		WithTxConfig(encodingConfig.TxConfig).
 		WithLegacyAmino(encodingConfig.Amino).
 		WithInput(os.Stdin).
-		WithAccountRetriever(types.AccountRetriever{}).
+		WithAccountRetriever(authtypes.AccountRetriever{}).
 		WithBroadcastMode(flags.BroadcastBlock).
 		WithHomeDir(app.DefaultNodeHome).
 		WithViper("DESMOS")
@@ -109,7 +112,8 @@ func initRootCmd(rootCmd *cobra.Command, encodingConfig params.EncodingConfig) {
 		rpc.StatusCommand(),
 		queryCommand(),
 		txCommand(),
-		GetSignCmd(),
+		sign.GetSignCmd(),
+		chainlink.GetCreateChainLinkJSON(),
 		keys.Commands(app.DefaultNodeHome),
 	)
 }